Runtime changes: - Some Array bugs fixed (especially sorting). - New Array methods: fromTypedArray, int indexer (for .NET code, not AS3) - Added ApplicationDomain.loadNativeClasses for importing multiple types - Fixed an issue with dynamic method calls on boxed primitive types - ASAny.AS_tryInvoke/AS_tryConstruct should now return false for null/undefined (instead of throwing an error) - Runtime weak and strict equality operators now correctly return false for NaN == NaN - ASArray class now has unit test coverage.
Compiler changes:
- Optimize multiple string concatenations (e.g. s1 + s2 + s3 + s4)
- Don't emit arguments array initialization code for a method with NEED_ARGUMENTS if the arguments array is not used.

+++ Mariana.AVM2/src/Core/AS/ASRegExp.cs
@@ -418,7 +418,7 @@ namespace Mariana.AVM2.Core {
 
                 for (int i = 0, n = groups.Count; i < n; i++) {
                     string groupValue = groups[i].Value;
-                    groupArray[(uint)i] = groupValue;
+                    groupArray[i] = groupValue;
 
                     if (i != 0 && m_groupNames != null && m_groupNames[i - 1] != null)
                         namedProps[m_groupNames[i - 1]] = groupValue;
